pyproject.toml
[tool.poetry]
name = "todo-react-django-ghostpost"
version = "0.1.0"
description = ""
authors = ["Kano Marvel <knmarvel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
flake9 = "^3.8.3"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
